Got a simple compiler working.
Had to introduce a very ugly hack into unwind, where it returns a value
if what it has is a fully evaluated number.  I am less than pleased with
this; for one, it isn't going to work if we want to strictly evaluate
something other than a number.  Also, there's an unnecessary clone
happening.

On the bright side, yay, something works.